Rename docker.rs to container.rs
